Fix compilation of units for DJGPP.
This doesn't fix running the units, several of which have issues,
but at least they build now without warnings.

+++ unit/audio/mixer.cpp
@@ -218,7 +218,7 @@ public:
         size_t pos = render.size();
         render.resize(pos + dest.size());
         for(size_t i = 0; i < dest.size(); i++)
-          render[pos + i] = Unit::clamp(dest[i], -32768, 32767);
+          render[pos + i] = Unit::clamp((int)dest[i], -32768, 32767);
       }
     }
